extract MetricNames to public classes
