+++ resilience4j-micrometer/src/main/java/io/github/resilience4j/micrometer/tagged/AbstractBulkheadMetrics.java
@@ -30,8 +30,13 @@ import static java.util.Objects.requireNonNull;
 
 abstract class AbstractBulkheadMetrics extends AbstractMetrics {
 
-    protected final MetricNames names;
+    protected final BulkheadMetricNames names;
 
+    protected AbstractBulkheadMetrics(BulkheadMetricNames names) {
+        this.names = requireNonNull(names);
+    }
+
+    @Deprecated
     protected AbstractBulkheadMetrics(MetricNames names) {
         this.names = requireNonNull(names);
     }
@@ -64,107 +69,6 @@ abstract class AbstractBulkheadMetrics extends AbstractMetrics {
 
     }
 
-    /**
-     * Defines possible configuration for metric names.
-     */
-    public static class MetricNames {
-
-        private static final String DEFAULT_PREFIX = "resilience4j.bulkhead";
-
-        public static final String DEFAULT_BULKHEAD_AVAILABLE_CONCURRENT_CALLS_METRIC_NAME =
-            DEFAULT_PREFIX + ".available.concurrent.calls";
-        public static final String DEFAULT_BULKHEAD_MAX_ALLOWED_CONCURRENT_CALLS_METRIC_NAME =
-            DEFAULT_PREFIX + ".max.allowed.concurrent.calls";
-        private String availableConcurrentCallsMetricName = DEFAULT_BULKHEAD_AVAILABLE_CONCURRENT_CALLS_METRIC_NAME;
-        private String maxAllowedConcurrentCallsMetricName = DEFAULT_BULKHEAD_MAX_ALLOWED_CONCURRENT_CALLS_METRIC_NAME;
-
-        private MetricNames() {
-        }
-
-        /**
-         * Returns a builder for creating custom metric names. Note that names have default values,
-         * so only desired metrics can be renamed.
-         *
-         * @return The builder.
-         */
-        public static Builder custom() {
-            return new Builder();
-        }
-
-        /**
-         * Returns default metric names.
-         *
-         * @return The default {@link TaggedBulkheadMetrics.MetricNames} instance.
-         */
-        public static MetricNames ofDefaults() {
-            return new MetricNames();
-        }
-
-        /**
-         * Returns the metric name for bulkhead concurrent calls, defaults to {@value
-         * DEFAULT_BULKHEAD_AVAILABLE_CONCURRENT_CALLS_METRIC_NAME}.
-         *
-         * @return The available concurrent calls metric name.
-         */
-        public String getAvailableConcurrentCallsMetricName() {
-            return availableConcurrentCallsMetricName;
-        }
-
-        /**
-         * Returns the metric name for bulkhead max available concurrent calls, defaults to {@value
-         * DEFAULT_BULKHEAD_MAX_ALLOWED_CONCURRENT_CALLS_METRIC_NAME}.
-         *
-         * @return The max allowed concurrent calls metric name.
-         */
-        public String getMaxAllowedConcurrentCallsMetricName() {
-            return maxAllowedConcurrentCallsMetricName;
-        }
-
-        /**
-         * Helps building custom instance of {@link TaggedBulkheadMetrics.MetricNames}.
-         */
-        public static class Builder {
-
-            private final MetricNames metricNames = new MetricNames();
-
-            /**
-             * Overrides the default metric name {@value TaggedBulkheadMetrics.MetricNames#DEFAULT_BULKHEAD_AVAILABLE_CONCURRENT_CALLS_METRIC_NAME}
-             * with a given one.
-             *
-             * @param availableConcurrentCallsMetricName The available concurrent calls metric
-             *                                           name.
-             * @return The builder.
-             */
-            public Builder availableConcurrentCallsMetricName(
-                String availableConcurrentCallsMetricName) {
-                metricNames.availableConcurrentCallsMetricName = requireNonNull(
-                    availableConcurrentCallsMetricName);
-                return this;
-            }
-
-            /**
-             * Overrides the default metric name {@value TaggedBulkheadMetrics.MetricNames#DEFAULT_BULKHEAD_MAX_ALLOWED_CONCURRENT_CALLS_METRIC_NAME}
-             * with a given one.
-             *
-             * @param maxAllowedConcurrentCallsMetricName The max allowed concurrent calls metric
-             *                                            name.
-             * @return The builder.
-             */
-            public Builder maxAllowedConcurrentCallsMetricName(
-                String maxAllowedConcurrentCallsMetricName) {
-                metricNames.maxAllowedConcurrentCallsMetricName = requireNonNull(
-                    maxAllowedConcurrentCallsMetricName);
-                return this;
-            }
-
-            /**
-             * Builds {@link TaggedBulkheadMetrics.MetricNames} instance.
-             *
-             * @return The built {@link TaggedBulkheadMetrics.MetricNames} instance.
-             */
-            public MetricNames build() {
-                return metricNames;
-            }
-        }
-    }
+    @Deprecated
+    public static class MetricNames extends BulkheadMetricNames { }
 }
