+++ resilience4j-micrometer/src/main/java/io/github/resilience4j/micrometer/tagged/AbstractTimeLimiterMetrics.java
@@ -34,8 +34,13 @@ abstract class AbstractTimeLimiterMetrics extends AbstractMetrics {
     private static final String KIND_SUCCESSFUL = "successful";
     private static final String KIND_TIMEOUT = "timeout";
 
-    protected final MetricNames names;
+    protected final TimeLimiterMetricNames names;
 
+    protected AbstractTimeLimiterMetrics(TimeLimiterMetricNames names) {
+        this.names = requireNonNull(names);
+    }
+
+    @Deprecated
     protected AbstractTimeLimiterMetrics(MetricNames names) {
         this.names = requireNonNull(names);
     }
@@ -77,69 +82,7 @@ abstract class AbstractTimeLimiterMetrics extends AbstractMetrics {
         meterIdMap.put(timeLimiter.getName(), new HashSet<>(ids));
     }
 
-    public static class MetricNames {
-
-        private static final String DEFAULT_PREFIX = "resilience4j.timelimiter";
-        public static final String DEFAULT_TIME_LIMITER_CALLS = DEFAULT_PREFIX + ".calls";
-
-        private String callsMetricName = DEFAULT_TIME_LIMITER_CALLS;
-
-        /**
-         * Returns a builder for creating custom metric names. Note that names have default values,
-         * so only desired metrics can be renamed.
-         *
-         * @return The builder.
-         */
-        public static Builder custom() {
-            return new Builder();
-        }
-
-        /**
-         * Returns default metric names.
-         *
-         * @return The default {@link MetricNames} instance.
-         */
-        public static MetricNames ofDefaults() {
-            return new MetricNames();
-        }
-
-        /**
-         * Returns the metric name for circuit breaker calls, defaults to {@value
-         * DEFAULT_TIME_LIMITER_CALLS}.
-         *
-         * @return The circuit breaker calls metric name.
-         */
-        public String getCallsMetricName() {
-            return callsMetricName;
-        }
-
-        /**
-         * Helps building custom instance of {@link MetricNames}.
-         */
-        public static class Builder {
-
-            private final MetricNames metricNames = new MetricNames();
-
-            /**
-             * Overrides the default metric name {@value MetricNames#DEFAULT_TIME_LIMITER_CALLS}
-             * with a given one.
-             *
-             * @param callsMetricName The calls metric name.
-             * @return The builder.
-             */
-            public Builder callsMetricName(String callsMetricName) {
-                metricNames.callsMetricName = requireNonNull(callsMetricName);
-                return this;
-            }
-
-            /**
-             * Builds {@link MetricNames} instance.
-             *
-             * @return The built {@link MetricNames} instance.
-             */
-            public MetricNames build() {
-                return metricNames;
-            }
-        }
+    @Deprecated
+    public static class MetricNames extends TimeLimiterMetricNames {
     }
 }
