+++ resilience4j-micrometer/src/test/java/io/github/resilience4j/micrometer/tagged/TaggedThreadPoolBulkheadMetricsPublisherTest.java
@@ -29,8 +29,8 @@ import org.junit.Test;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static io.github.resilience4j.micrometer.tagged.AbstractThreadPoolBulkheadMetrics.MetricNames.*;
 import static io.github.resilience4j.micrometer.tagged.MetricsTestHelper.findMeterByNamesTag;
+import static io.github.resilience4j.micrometer.tagged.ThreadPoolBulkheadMetricNames.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class TaggedThreadPoolBulkheadMetricsPublisherTest {
@@ -163,7 +163,7 @@ public class TaggedThreadPoolBulkheadMetricsPublisherTest {
         MeterRegistry meterRegistry = new SimpleMeterRegistry();
         TaggedThreadPoolBulkheadMetricsPublisher taggedBulkheadMetricsPublisher =
             new TaggedThreadPoolBulkheadMetricsPublisher(
-                TaggedThreadPoolBulkheadMetricsPublisher.MetricNames.custom()
+                ThreadPoolBulkheadMetricNames.custom()
                     .maxThreadPoolSizeMetricName("custom.max.thread.pool.size")
                     .coreThreadPoolSizeMetricName("custom.core.thread.pool.size")
                     .build(), meterRegistry);
